requirements.txt
alembic==1.4.2
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.10
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0.7
django-request==1.5.5
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
keyring==18.0.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.2.2
monotonic==1.5
MyApplication==0.1.0
netifaces==0.10.4
numpy==1.19.0
oauthlib==3.1.0
olefile==0.46
pandas==1.0.5
paramiko==2.6.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==0.23.1
scipy==1.5.0
seaborn==0.10.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.18
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
systemd-python==234
Tempita==0.5.2
threadpoolctl==2.1.0
torch==1.5.1
torchvision==0.6.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.25
wadllib==1.3.3
Werkzeug==1.0.1
WTForms==2.3.1
xkit==0.0.0
